requirements.txt
tornado~=6.5.1
psycopg2~=2.9.10
nltk~=3.9.1
scikit-learn~=1.7.0
rank_bm25~=0.2.2
sentence-transformers~=4.1.0
gensim~=4.3.3
psycopg2-binary~=2.9.10
numpy~=1.26.4
joblib~=1.5.1
torch~=2.7.1
transformers~=4.52.4
scipy~=1.13.1
faiss-cpu~=1.11.0